Tested RDMNET-80, found & fixed issues
The issue detected was that it was not possible to set the static IPv4 of a device. At first the device would respond with a NACK because it wasn't getting the right scope slot, due to a controller bug. After that was fixed, the port number was not being sent correctly, so that was fixed as well. Refer to RDMnetNetworkModel.cpp. A new SCR will not be created since this is a part of RDMNET-80.