Fixed RDMNET-92
The issue was that BrokerConnection::connect() was being called multiple times on the same connection, after already being connected. This would result in a call to rdmnet_init_dynamic_uid_request(&local_uid_, 0x6574) in BrokerConnection::runConnectStateMachine(), leaving the local UID ID at 0, and thus causing communication issues down the road.

An extra condition was added in this function so that it would only start the connect state machine thread if the BrokerConnection::connected_ == false. This fixed the issue.